+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/test_mgmt.py
@@ -103,7 +103,7 @@ class TestManagement(FormRecognizerTest):
         models_list = client.list_custom_models()
         for model in models_list:
             self.assertIsNotNone(model.model_id)
-            self.assertEqual(model.status, "ready")
+            self.assertIsNotNone(model.status)
             self.assertIsNotNone(model.created_on)
             self.assertIsNotNone(model.last_modified)
 
@@ -138,7 +138,7 @@ class TestManagement(FormRecognizerTest):
         models_list = client.list_custom_models()
         for model in models_list:
             self.assertIsNotNone(model.model_id)
-            self.assertEqual(model.status, "ready")
+            self.assertIsNotNone(model.status)
             self.assertIsNotNone(model.created_on)
             self.assertIsNotNone(model.last_modified)