+++ sdk/formrecognizer/azure-ai-formrecognizer/CHANGELOG.md
@@ -21,6 +21,7 @@ For recognize receipt methods, parameters have been renamed to `receipt` and `re
 
 **New features**
 
+- Support to copy a custom model from one Form Recognizer resource to another
 - Authentication using `azure-identity` credentials now supported
   - see the [Azure Identity documentation](https://github.com/Azure/azure-sdk-for-python/blob/master/sdk/identity/azure-identity/README.md) for more information